Replaces all 'unsigned long long' statements with a typedef at the top of the tools.h file (it is included in every c++ file) and changed all the 'unsigned long long' statements in the rest of the code accordingly. Switched all functions dealing with hash from using 8 to sizeof(index_type).

+++ python/seqdb2.pyx
@@ -10,7 +10,7 @@ cdef extern from "dbread.h":
     ctypedef char * (*getTypeByIndex)(unsigned)
     ctypedef void (*getRecord)(unsigned long long)
     ctypedef void (*clear)()
-    ctypedef unsigned long long (*getSize)()
+    ctypedef unsigned long long(*getSize)()
     ctypedef unsigned (*getTypesize)()
     ctypedef char * (*getTypekey)(unsigned)
     ctypedef void (*getHashRecord)(char*, unsigned)
